refactor(struct): extract EPUB reader-facing fallback tests
